+++ grakn-client/src/main/java/ai/grakn/remote/GrpcClient.java
@@ -21,18 +21,28 @@ package ai.grakn.remote;
 import ai.grakn.GraknTx;
 import ai.grakn.GraknTxType;
 import ai.grakn.Keyspace;
+import ai.grakn.concept.Concept;
+import ai.grakn.concept.ConceptId;
+import ai.grakn.graql.Graql;
 import ai.grakn.graql.Query;
+import ai.grakn.graql.Var;
+import ai.grakn.graql.admin.Answer;
+import ai.grakn.graql.internal.query.QueryAnswer;
 import ai.grakn.grpc.GrpcUtil;
 import ai.grakn.grpc.GrpcUtil.ErrorType;
 import ai.grakn.grpc.TxGrpcCommunicator;
 import ai.grakn.grpc.TxGrpcCommunicator.Response;
+import ai.grakn.remote.concept.RemoteConcepts;
 import ai.grakn.rpc.generated.GraknGrpc;
+import ai.grakn.rpc.generated.GraknOuterClass;
 import ai.grakn.rpc.generated.GraknOuterClass.TxResponse;
 import ai.grakn.util.CommonUtil;
 import com.google.common.collect.AbstractIterator;
+import com.google.common.collect.ImmutableMap;
 import io.grpc.Metadata;
 import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
+import mjson.Json;
 
 import javax.annotation.Nullable;
 import java.util.Iterator;
@@ -83,7 +93,7 @@ final class GrpcClient implements AutoCloseable {
 
                 switch (response.getResponseCase()) {
                     case QUERYRESULT:
-                        return GrpcUtil.getQueryResult(tx, response.getQueryResult());
+                        return convert(tx, response.getQueryResult());
                     case DONE:
                         return endOfData();
                     default:
@@ -138,6 +148,56 @@ final class GrpcClient implements AutoCloseable {
             return errorType.toException(message);
         } else {
             return error;
+        }
+    }
+
+    private static Object convert(GraknTx tx, GraknOuterClass.QueryResult queryResult) {
+        switch (queryResult.getQueryResultCase()) {
+            case ANSWER:
+                return convert(tx, queryResult.getAnswer());
+            case OTHERRESULT:
+                return Json.read(queryResult.getOtherResult()).getValue();
+            default:
+            case QUERYRESULT_NOT_SET:
+                throw new IllegalArgumentException("Unexpected " + queryResult);
+        }
+    }
+
+    private static Answer convert(GraknTx tx, GraknOuterClass.Answer answer) {
+        ImmutableMap.Builder<Var, Concept> map = ImmutableMap.builder();
+
+        answer.getAnswerMap().forEach((grpcVar, grpcConcept) -> {
+            map.put(Graql.var(grpcVar), convert(tx, grpcConcept));
+        });
+
+        return new QueryAnswer(map.build());
+    }
+
+    private static Concept convert(GraknTx tx, GraknOuterClass.Concept concept) {
+        ConceptId id = ConceptId.of(concept.getId());
+
+        switch (concept.getBaseType()) {
+            case Entity:
+                return RemoteConcepts.createEntity(tx, id);
+            case Relationship:
+                return RemoteConcepts.createRelationship(tx, id);
+            case Attribute:
+                return RemoteConcepts.createAttribute(tx, id);
+            case EntityType:
+                return RemoteConcepts.createEntityType(tx, id);
+            case RelationshipType:
+                return RemoteConcepts.createRelationshipType(tx, id);
+            case AttributeType:
+                return RemoteConcepts.createAttributeType(tx, id);
+            case Role:
+                return RemoteConcepts.createRole(tx, id);
+            case Rule:
+                return RemoteConcepts.createRule(tx, id);
+            case MetaType:
+                return RemoteConcepts.createMetaType(tx, id);
+            default:
+            case UNRECOGNIZED:
+                throw new IllegalArgumentException("Unrecognised " + concept);
         }
     }
 }
