+++ grakn-client/src/main/java/ai/grakn/remote/concept/RemoteConcepts.java
@@ -16,7 +16,7 @@
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
  */
 
-package ai.grakn.grpc.concept;
+package ai.grakn.remote.concept;
 
 import ai.grakn.GraknTx;
 import ai.grakn.concept.Attribute;
